requirements-dev.txt
black==26.1.0
cfgv==3.5.0
click==8.3.1
distlib==0.4.0
filelock==3.20.3
flake8==7.3.0
identify==2.6.16
isort==7.0.0
librt==0.7.8
mccabe==0.7.0
mypy==1.19.1
mypy_extensions==1.1.0
nodeenv==1.10.0
packaging==26.0
pathspec==1.0.4
platformdirs==4.5.1
pre_commit==4.5.1
psutil==7.2.2
pycodestyle==2.14.0
pyflakes==3.4.0
pytokens==0.4.1
PyYAML==6.0.3
types-psutil==7.2.2.20260130
types-PyYAML==6.0.12.20250915
typing_extensions==4.15.0
virtualenv==20.36.1
